Live script 2.9.3, ArduinoJson fix
pio.ini
- Live Script 2.9.1.1 to 2.9.3

SysModPins
- board var: send only if clients, don't call sendDataWS but call sendBuffer

SysModWeb
- sendBuffer: add parameter lossless
- serializeState, serializeInfo: Fix for newer arduinoJson: deserializeJson only accepts jsonVariants
- sendResponseObject: don't call sendDataWS but call sendBuffer

UserModLive
- replace run with compile and add execute(Background)Task and taskExists
- script var: compile if not taskExists, remove kill()
- update onSetValue of scripts table columns
- compile: replace executeAsTask by addExe
- refactor kill (kill + deleteExe)

+++ src/Sys/SysModModel.h
@@ -51,7 +51,7 @@ struct Coord3D {
   bool operator<(Coord3D rhs) {
     return x < rhs.x && y < rhs.y && z < rhs.z;
   }
-  bool operator>=(uint16_t rhs) {
+  bool operator>=(int rhs) {
     return x >= rhs && y >= rhs && z >= rhs;
   }
